requirements.txt
Pillow==8.1.2
Pillow==10.0.1
pyexcel==0.7.0
